+++ src/azure-cli/HISTORY.rst
@@ -11,6 +11,9 @@ Release History
 
 * Support import/export of keyvault references from/to appservice
 * Support import/export of all labels from appconfig to appconfig
+* Validate key and feature names before setting and importing
+* Expose sku modification for configuration store.
+* Add command group for managed identity.
 
 **AppService**
 
@@ -20,6 +23,10 @@ Release History
 
 * Fix issue #11658: `az group export` command does not support `--query` and `--output` parameters
 * Fix issue #10279: The exit code of `az group deployment validate` is 0 when the verification fails
+
+**IoT**
+
+* Deprecated 'IoT hub Job' commands.
 
 **IoT Central**
 
@@ -385,6 +392,7 @@ Release History
 * Change default node version on function apps to ~10 for Windows
 * Add --runtime-version property to `az functionapp create`
 * az appservice vnet-integration add: Fixed so that subnet delegation is case insensitive and delegating subnets does not overwrite previous data.
+
 
 **ARM**